Some BC++ issues
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@5129 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/html/m_fonts.cpp
@@ -55,7 +55,7 @@ TAG_HANDLER_BEGIN(FONT, "FONT")
 
         if (tag.HasParam(wxT("SIZE"))) {
 	    long tmp = 0;
-            wxChar c = tag.GetParam(wxT("SIZE"))[0];
+            wxChar c = tag.GetParam(wxT("SIZE"))[(unsigned int) 0];
             if (tag.ScanParam(wxT("SIZE"), wxT("%li"), &tmp) == 1) {
                 if (c == '+' || c == '-')
                     m_WParser -> SetFontSize(oldsize+tmp);
